no-mistakes(review): Recheck queue eligibility before branch wake handling
